object_detection/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "urban-det"
version = "0.1.0"
description = "SOTA object detection for urban traffic scenes"
requires-python = ">=3.11"
dependencies = [
    # Core ML
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "pytorch-lightning>=2.3.0",
    "torchmetrics>=1.4.0",
    # Model zoo
    "transformers>=4.41.0",
    "timm>=1.0.3",
    "einops>=0.8.0",
    # Data
    "pycocotools>=2.0.7",
    "albumentations>=1.4.0",
    "opencv-python-headless>=4.9.0",
    "Pillow>=10.3.0",
    # Config
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    # Experiment tracking
    "mlflow>=2.13.0",
    "wandb>=0.17.0",
    # Deployment
    "onnx>=1.16.0",
    "onnxruntime-gpu>=1.18.0",
    "onnxsim>=0.4.36",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Monitoring
    "prometheus-client>=0.20.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    # Utilities
    "scipy>=1.13.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
    "pydantic>=2.7.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
tensorrt = [
    "tensorrt>=10.0.0",
    "cuda-python>=12.4.0",
    "pycuda>=2024.1",
]
ros2 = [
    "rclpy",
    "sensor-msgs",
    "vision-msgs",
]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-Pillow",
    "httpx>=0.27.0",
]

[project.scripts]
urban-det-train = "urban_det.scripts.train:main"
urban-det-eval = "urban_det.scripts.evaluate:main"
urban-det-export = "urban_det.scripts.export_model:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=urban_det --cov-report=term-missing -q"
asyncio_mode = "auto"
